adjusted integer types in renderer

+++ ST_engine/ST_renderer/include/renderer_sdl.hpp
@@ -32,21 +32,21 @@ namespace ST {
 
         void draw_background(size_t arg);
 
-        void draw_overlay(size_t arg, int sprite, int sprite_num);
+        void draw_overlay(size_t arg, uint8_t sprite, uint8_t sprite_num);
 
-        void draw_texture(size_t arg, int x, int y);
+        void draw_texture(size_t arg, int32_t x, int32_t y);
 
-        void draw_texture_scaled(size_t arg, int x, int y, float scale_x, float scale_y);
+        void draw_texture_scaled(size_t arg, int32_t x, int32_t y, float scale_x, float scale_y);
 
-        void draw_rectangle(int x, int y, int w, int h, SDL_Color color);
+        void draw_rectangle(int32_t x, int32_t y, int32_t w, int32_t h, SDL_Color color);
 
-        void draw_rectangle_filled(int x, int y, int w, int h, SDL_Color color);
+        void draw_rectangle_filled(int32_t x, int32_t y, int32_t w, int32_t h, SDL_Color color);
 
-        void draw_sprite(size_t arg, int x, int y, int sprite, int animation, int animation_num, int sprite_num);
+        void draw_sprite(size_t arg, int32_t x, int32_t y, uint8_t sprite, uint8_t animation, uint8_t animation_num, uint8_t sprite_num);
 
-        void draw_sprite_scaled(size_t arg, int x, int y, int sprite, int animation, int animation_num, int sprite_num, float scale_x, float scale_y);
+        void draw_sprite_scaled(size_t arg, int32_t x, int32_t y, uint8_t sprite, uint8_t animation, uint8_t animation_num, uint8_t sprite_num, float scale_x, float scale_y);
         
-        uint16_t draw_text(const std::string& arg, const std::string& arg2, int x, int y, SDL_Color color_font , uint8_t size, int flag);
+        uint16_t draw_text(const std::string& arg, const std::string& arg2, int32_t x, int32_t y, SDL_Color color_font , uint8_t size, int8_t flag);
 
         void upload_surfaces(ska::bytell_hash_map<size_t, SDL_Surface *> *surfaces);
